Create function to generate custom grid
One of the constraints of the assignment is that the grid has to be
dynamically generated based on the grid size chosen by the user. The
default size is 16x16. The function added here requires a grid size to
be provided and then generates a custom grid based on that number.
(Example: 16x16, 8x8, 4x4)